+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/intellij/link/LinkConfig.java
@@ -13,14 +13,14 @@ import lombok.ToString;
 @Getter
 @Setter
 @ToString
-public class LinkComposite<S extends BaseLinkConfig, T extends BaseLinkConfig> {
+public class LinkConfig<R extends BaseResourceConfig, T extends BaseResourceConfig> {
 
-    private S service;
+    private R resource;
     private T module;
     private String envPrefix;
 
-    public LinkComposite(S source, T target) {
-        this.service = source;
+    public LinkConfig(R resource, T target) {
+        this.resource = resource;
         this.module = target;
     }