+++ PluginsAndFeatures/azure-toolkit-for-intellij/resources/whatsnew/whatsnew.md
@@ -1,5 +1,15 @@
-<!-- Version: 3.47.0 -->
+<!-- Version: 3.50.0 -->
 # What's new in Azure Toolkit for IntelliJ
+
+## 3.50.0
+
+### Added
+<img src="https://user-images.githubusercontent.com/19339116/109937625-11666400-7d0a-11eb-9850-82a62d65f3fa.gif" width="840" height="525" />
+
+- Development workflow for Azure Database for MySQL
+  - Connect Azure Database for MySQL Server to local project from Azure Explorer or application.properties file
+  - Automatically inject datasource connection properties into runtime environment for local run
+  - Publish Azure Web App with datasource connection properties in application settings
 
 ## 3.49.0